% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{choose_directory}
\alias{choose_directory}
\title{Choose the project's directory.}
\usage{
choose_directory(caption = "Select data directory")
}
\arguments{
\item{caption}{Title shown in the folder-selection dialog (default
\code{'Select data directory'}).}
}
\value{
The user's selected directory.
}
\description{
From the easycsv package.
}
