% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{clusterCols}
\alias{clusterCols}
\title{Set cluster colors.}
\usage{
clusterCols(n, type = "rcmap")
}
\arguments{
\item{n}{Number of colors.}

\item{type}{Either rcmap (default) or rainbow.}
}
\description{
Either use the RCMap scheme which has 21 colors, or the rainbow function.
}
