% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{get_cmat}
\alias{get_cmat}
\title{Build co-membership matrices for a range of cluster solutions.}
\usage{
get_cmat(dists, M, disttype, max_nc, clustMethod = "ward.D2")
}
\arguments{
\item{dists}{An MxM distance matrix (e.g. Euclidean or hyperbolic 2-D distances).}

\item{M}{Number of statements (rows/columns of \code{dists}).}

\item{disttype}{Character label for the distance type (e.g. \code{"Euclidean"}
or \code{"Hyperbolic"}). Currently used for labelling only; does not alter
the computation.}

\item{max_nc}{Maximum number of clusters to consider.}

\item{clustMethod}{Hierarchical clustering method passed to \code{hclust}
(default \code{"ward.D2"}).}
}
\value{
A list of length \code{max_nc - 1}. Element \code{[[k-1]]} is an
  MxM 0/1 co-membership matrix for the k-cluster solution (diagonal is 0).
}
\description{
Runs hierarchical clustering on a distance matrix and, for each number of
clusters k from 2 to \code{max_nc}, builds an MxM binary co-membership
matrix where entry (i,j) is 1 if statements i and j are in the same cluster.
}
