% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{getOS}
\alias{getOS}
\title{Detect the current operating system.}
\usage{
getOS()
}
\value{
A character string: \code{"windows"}, \code{"Linux"}, or
  \code{"MacOSX"}.
}
\description{
Detect the current operating system.
}
