% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{MPindex}
\alias{MPindex}
\title{Calculate a misplacement index.}
\usage{
MPindex(D, d, cst = 1, func = median, ...)
}
\arguments{
\item{D}{dissimilarity matrix obtained for the pile sorting data.}

\item{d}{distance matrix obtained for points in the 2-D projection.}

\item{cst}{The power parameter for the misplacement index function (default=1).}

\item{func}{the function applied to the pair of matrices, d and D. Default=median.}

\item{...}{additional parameters to be passed to func.}
}
\value{
A vector of size nrow(D) with the bridging/anchoring metric.
}
\description{
Takes a dissimilarity matrix from a pile-sorting study and a distance matrix
obtained from the 2-D map (from multidimensional scaling) and returns an
index which is close to 0 for statements which are positioned close (in 2-D)
to other statements with which they put in the same pile often (anchors),
and a value close to 1 for points which are positioned in the 2-D map
far from other  statements with which were put in the same pile (bridges).
}
