% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{raterReport}
\alias{raterReport}
\title{Print a summary report about the raters.}
\usage{
raterReport()
}
\description{
Prints a statistical summary of all demographic variables for the raters.
Pauses for user input before returning to the menu.
}
