% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{ratingSummary}
\alias{ratingSummary}
\title{Get ratings statistics (N, mean, sd, min, max) for each statement}
\usage{
ratingSummary(ratingsDat)
}
\arguments{
\item{ratingsDat}{A data frame where column 1 is \code{RaterID}, column 2 is
\code{StatementID} (will be coerced to factor), and columns 3 onward are
numeric rating variables.}
}
\value{
A matrix with summary statistics for all rating variables and all
  statements. Each block of 5 columns corresponds to one rating variable
  (N, Mean, SD, Min, Max). The final row contains grand-total summary values
  across all statements.
}
\description{
Raters provide Likert scale ratings for k criteria (e.g. feasibility). This function returns summary statistics for each statement.
}
