% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{read_statements}
\alias{read_statements}
\title{Read statements file with basic validation and normalization.}
\usage{
read_statements(filename, enc = "UTF-8", sep = ",")
}
\arguments{
\item{filename}{Path to the statements CSV file.}

\item{enc}{File encoding (default UTF-8).}

\item{sep}{Field separator (default comma).}
}
\value{
A list with elements \code{issues} (NULL if none) and \code{cardNames} (data.frame)
}
\description{
The statements file should be a CSV with two columns (ID and statement text).
The function is forgiving about header names (it will rename the first two
columns to `StatementID` and `Statement`) and will normalize encodings and
whitespace. It validates that StatementID values are numeric and unique.
}
\examples{
\donttest{
read_statements('Statements.csv')
}
}
