% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{sorterReport}
\alias{sorterReport}
\title{Print a summary report about each sorter.}
\usage{
sorterReport()
}
\value{
Called for side effects. Returns \code{NULL} invisibly.
}
\description{
For each sorter, prints the number of cards sorted and the number of piles
created. Pauses for user input before returning to the menu.
}
