% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{sumsq}
\alias{sumsq}
\title{Calculate the sum of squares of a vector}
\usage{
sumsq(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
The sum of squares.
}
\description{
Calculate the sum of squares of a vector
}
