% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{clusterTUKEY}
\alias{clusterTUKEY}
\title{All pairwise comparisons with Tukey's method for multiple comparisons.}
\usage{
clusterTUKEY()
}
\description{
All pairwise comparisons with Tukey's method for multiple comparisons.
}
