% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{diskDist}
\alias{diskDist}
\title{Convert Euclidean distances in the unit disk to a hyperbolic space distance.}
\usage{
diskDist(x)
}
\arguments{
\item{x}{distance matrix obtained for points in 2-D.}
}
\value{
The hyperbolic distances.
}
\description{
Converts a Euclidean distance matrix between points in a unit 2-D disk into a
(Poincare) distance in a 2-D hyperbolic space.
}
