% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Add adjacency matrices from pile-sorting data and return a distance matrix}
\usage{
distanceMatrix(adjMats)
}
\arguments{
\item{adjMats}{a list of n SxS adjacency matrices, created by getAdjMatrices.}
}
\value{
An SxS distance matrix.
}
\description{
n SxS adjacency matrices are added to create a similarity matrix, which is
then converted into a distance matrix in the S-dimensional Euclidean space.
}
