% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{getAdjMatrices}
\alias{getAdjMatrices}
\title{Create adjacency matrices from pile-sorting data}
\usage{
getAdjMatrices(piledat, showWarnings = TRUE)
}
\arguments{
\item{piledat}{The pile sorting data.}

\item{showWarnings}{Print any potential problems in the pile-sorting data
(default=TRUE).}
}
\value{
A list of n SxS 0/1-matrices - one for each sorter.
}
\description{
N stakeholder sort S cards into piles, and this function converts this
data into n SxS adjacency matrices (one per sorter) in which cell i,j
contains 1 if user k put cards i and j in the same pile, and 0 otherwise.
The first column in the input file contains the sorters IDs, the
second column contains a name for the pile, given by the sorter
(this column may contain blanks), and the column 3,4,... contain
card numbers. If a user x created M piles, the file will contain
M rows with x in the first column.
Note that a card cannot be in more than one pile.
Sorters cannot put all the cards in one pile, but they can leave
cards unsorted.
}
