% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{initCMap}
\alias{initCMap}
\title{Read the card-sorting data files and generate the MDS data.}
\usage{
initCMap(dataDir)
}
\arguments{
\item{dataDir}{The project's directory.}
}
\value{
A list with 9 elements n.indiv=number of sorters, cardNames=the statements on the cards, adj.mat=the adjacency matrix, DS=the pairwise distances matrix in the original high-dimensional space, D2e=the Euclidean distance matrix in the 2-D space, D2h=the hyperbolic distance matrix in the 2-D space, x=the x coordinates in the MDS plot, y=the y coordinates, stress=the stress of the MDS plot.
}
\description{
Takes two input file names.
}
