% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{ratingSummary}
\alias{ratingSummary}
\title{Get ratings statistics (N, mean, sd, min, max) for each statement}
\usage{
ratingSummary(ratingsDat)
}
\arguments{
\item{ratingsDat}{The ratings data.}
}
\value{
A matrix with summary statistics for all rating variables and all statements.
}
\description{
Raters provide Likert scale ratings for k criteria (e.g. feasibility). This function returns summary statistics for each statement.
}
