% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{showBarPlot}
\alias{showBarPlot}
\title{Show average ratings by clusters as a barplot.}
\usage{
showBarPlot(metric = NULL)
}
\arguments{
\item{metric}{Euclidean or hyperbolic.}
}
\description{
Show average ratings by clusters as a barplot.
}
