% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{showClusterPlot}
\alias{showClusterPlot}
\title{Plot the clusters in the 2-D representation of the pile-sorting data}
\usage{
showClusterPlot(type = "rays", metric = NULL)
}
\arguments{
\item{type}{Rays, or polygon.}

\item{metric}{cols The metric to be used (Euclidean or hyperbolic.)}
}
\description{
Can show the clusters either as convex polygons, or as lines connected to the center of each cluster.
}
