% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{showDendrogram}
\alias{showDendrogram}
\title{Plot the clusters as a dendrogram or a phylogenic tree plot.}
\usage{
showDendrogram(phylo = FALSE)
}
\arguments{
\item{phylo}{Whether to show a dendrogram (a default) or a phylogenic tree.}
}
\description{
Plot the clusters as a dendrogram or a phylogenic tree plot.
}
