% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{showDotPlot}
\alias{showDotPlot}
\title{Show statement ratings as a dotchart.}
\usage{
showDotPlot(metric = NULL)
}
\arguments{
\item{metric}{Euclidean or hyperbolic.}
}
\description{
Show statement ratings as a dotchart.
}
