% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{showMDSPlot}
\alias{showMDSPlot}
\title{Plot the 2-D representation of the pile-sorting data}
\usage{
showMDSPlot(cols = "blue", metric = NULL)
}
\arguments{
\item{cols}{The colors of the data (default= all blue.)}

\item{metric}{The metric to be used (Euclidean or hyperbolic.)}
}
\description{
Plot the 2-D representation of the pile-sorting data
}
