% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{splitHalf}
\alias{splitHalf}
\title{Calculate reliability metrics.}
\usage{
splitHalf(
  adjMat,
  B = 10,
  disttype = "Hyperbolic",
  seed = 968421,
  plotit = FALSE
)
}
\arguments{
\item{adjMat}{pile-sorting data (n 0-1 matrices).}

\item{B}{the number of random splits (default=10).}

\item{disttype}{The distance metric to be used ("Hyperbolic", otherwise it's assumed to be Euclidean).}

\item{seed}{The random seed to be used.}

\item{plotit}{if TRUE, show a plot of the differences between the misplacement index between the two random halves.}
}
\value{
A list with two elements: reliability, which contains the mean correlation between the splits, the distance metric being used, and the function used to calculate the misplacement index.
}
\description{
Takes a dissimilarity matrix from a pile-sorting study and a distance matrix
obtained from the 2-D map (from multidimensional scaling) and returns an
index which is close to 0 for statements which are positioned close (in 2-D)
to other statements with which they put in the same pile often (anchors),
and a value close to 1 for points which are positioned in the 2-D map
far from other  statements with which were put in the same pile (bridges).
}
