% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMap.R
\name{toText}
\alias{toText}
\title{Convert menu choices to text}
\usage{
toText(varType, choice)
}
\arguments{
\item{varType}{The type of variable (e.g., "clusteringMethod", "distance")}

\item{choice}{The number of the selected menu item}
}
\description{
Convert menu choices to text
}
